Fixes issues #9 and #5
- Fixed all the deprecations
- Fixed an issue where the resizing of the views would not work
- Now requires 7.0+

+++ TabbediAd/TableViewController.m
@@ -99,7 +99,7 @@
     }
     
     // Configure the cell...
-    cell.textLabel.text = [NSString stringWithFormat:@"Cell %d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"Cell %ld", (long)indexPath.row];
     
     return cell;
 }
